Moving all counter plugins into a separate directory
