+++ src/components/performance_counters/memory/mem_counter_windows.cpp
