+++ src/components/performance_counters/papi/CMakeLists.txt
@@ -5,7 +5,7 @@
 # Distributed under the Boost Software License, Version 1.0. (See accompanying
 # file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-set(root "${PROJECT_SOURCE_DIR}/hpx/components/papi_counters")
+set(root "${PROJECT_SOURCE_DIR}/hpx/components/performance_counters/papi_counters")
 
 ################################################################################
 # PAPI event counters
