+++ src/components/performance_counters/papi/server/papi.cpp
@@ -14,8 +14,8 @@
 #include <hpx/runtime/actions/continuation.hpp>
 #include <hpx/util/thread_mapper.hpp>
 #include <hpx/util/high_resolution_clock.hpp>
-#include <hpx/components/papi/server/papi.hpp>
-#include <hpx/components/papi/util/papi.hpp>
+#include <hpx/components/performance_counters/papi/server/papi.hpp>
+#include <hpx/components/performance_counters/papi/util/papi.hpp>
 #include <hpx/exception.hpp>
 
 #include <functional>
