+++ src/components/performance_counters/papi/util/papi.cpp
@@ -11,7 +11,7 @@
 #include <hpx/config/asio.hpp>
 #include <hpx/util/parse_command_line.hpp>
 #include <hpx/util/thread_mapper.hpp>
-#include <hpx/components/papi/util/papi.hpp>
+#include <hpx/components/performance_counters/papi/util/papi.hpp>
 
 #include <boost/format.hpp>
 #include <boost/asio/ip/host_name.hpp>
